requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
attrs==21.4.0
backcall==0.2.0
beautifulsoup4==4.10.0
bleach==4.1.0
cffi==1.15.0
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
executing==0.8.3
fonttools==4.31.1
imageio==2.16.1
imutils==0.5.4
ipykernel==6.9.2
ipython==8.1.1
ipython-genutils==0.2.0
ipywidgets==7.7.0
jedi==0.18.1
Jinja2==3.0.3
jsonschema==4.4.0
jupyter==1.0.0
jupyter-client==7.1.2
jupyter-console==6.4.3
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.1.0
jupyterthemes==0.20.0
kiwisolver==1.4.0
lesscpy==0.15.0
MarkupSafe==2.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==0.8.4
MouseInfo==0.1.3
mss==6.1.0
nbclient==0.5.13
nbconvert==6.4.4
nbformat==5.2.0
nest-asyncio==1.5.4
networkx==2.7.1
notebook==6.4.10
numpy==1.22.3
opencv-python==4.5.5.64
packaging==21.3
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
ply==3.11
prometheus-client==0.13.1
prompt-toolkit==3.0.28
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
PyAutoGUI==0.9.53
pycparser==2.21
PyGetWindow==0.0.9
Pygments==2.11.2
PyMsgBox==1.0.9
pyparsing==3.0.7
pyperclip==1.8.2
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.10.1
PyRect==0.2.0
pyrsistent==0.18.1
PyScreeze==0.1.28
python-dateutil==2.8.2
python3-xlib==0.15
pytweening==1.0.4
PyWavelets==1.3.0
pyzmq==22.3.0
qtconsole==5.2.2
QtPy==2.0.1
scikit-image==0.19.2
scikit-video==1.1.11
scipy==1.8.0
Send2Trash==1.8.0
six==1.16.0
soupsieve==2.3.1
stack-data==0.2.0
terminado==0.13.3
testpath==0.6.0
tifffile==2022.3.16
tornado==6.1
traitlets==5.1.1
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.6.0
